requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
numpy>=1.23.0
matplotlib>=3.6.0
scikit-learn>=1.2.0
opencv-python>=4.7.0
seaborn>=0.12.0
flask>=2.3.0
grad-cam>=1.4.6
tqdm>=4.65.0